Removed global functionality out of masthead
