+++ client/galaxy/scripts/onload/globalInits/index.js
@@ -19,6 +19,7 @@ import { onloadWebhooks } from "./onloadWebhooks";
 import { replace_big_select_inputs } from "./replace_big_select_inputs";
 import { make_popup_menus } from "ui/popupmenu";
 import { initTours } from "./initTours";
+import { initModals } from "./initModals";
 
 export function globalInits() {
     prependInitialization(
@@ -26,6 +27,7 @@ export function globalInits() {
         initSentry,
         addIframeClass,
         adjustIframeLinks,
+        initModals,
         initTooltips,
         init_refresh_on_change,
         () => replace_big_select_inputs(20, 1500),
